+++ internal/image/process_test.go
@@ -8,7 +8,7 @@ import (
 	"image/png"
 	"testing"
 
-	"github.com/alexwitherspoon/aviationwx-bridge/internal/config"
+	"github.com/alexwitherspoon/AviationWX.org-Bridge/internal/config"
 )
 
 // createTestJPEG creates a test JPEG image with the given dimensions