+++ internal/web/server.go
@@ -12,8 +12,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/alexwitherspoon/aviationwx-bridge/internal/config"
-	"github.com/alexwitherspoon/aviationwx-bridge/internal/logger"
+	"github.com/alexwitherspoon/AviationWX.org-Bridge/internal/config"
+	"github.com/alexwitherspoon/AviationWX.org-Bridge/internal/logger"
 )
 
 //go:embed static/*